Fix System.Diagnostics.Process tests for uapaot runs
