+++ src/Common/src/Interop/Windows/user32/Interop.PostMessage.cs
@@ -9,7 +9,7 @@ internal partial class Interop
 {
     internal partial class User32
     {
-        [DllImport(Libraries.User32)]
+        [DllImport(Libraries.User32, EntryPoint = "PostMessageW")]
         public static extern int PostMessage(IntPtr hwnd, int msg, IntPtr wparam, IntPtr lparam);
     }
 }
