+++ src/Common/src/Interop/Windows/user32/Interop.SendMessageTimeout.cs
@@ -9,7 +9,7 @@ internal partial class Interop
 {
     internal partial class User32
     {
-	    [DllImport(Libraries.User32)]
+	    [DllImport(Libraries.User32, EntryPoint = "SendMessageTimeoutW")]
         public static extern IntPtr SendMessageTimeout(IntPtr hWnd, int msg, IntPtr wParam, IntPtr lParam, int flags, int timeout, out IntPtr pdwResult);
     }
 }
