+++ src/System.Diagnostics.Process/src/PinvokeAnalyzerExceptionList.analyzerdata
@@ -24,11 +24,11 @@ ntdll.dll!NtQueryInformationProcess
 ntdll.dll!NtQuerySystemInformation
 user32.dll!EnumWindows
 user32.dll!GetWindow
-user32.dll!GetWindowLong
-user32.dll!GetWindowText
-user32.dll!GetWindowTextLength
+user32.dll!GetWindowLongW
+user32.dll!GetWindowTextW
+user32.dll!GetWindowTextLengthW
 user32.dll!GetWindowThreadProcessId
 user32.dll!IsWindowVisible
-user32.dll!PostMessage
-user32.dll!SendMessageTimeout
+user32.dll!PostMessageW
+user32.dll!SendMessageTimeoutW
 user32.dll!WaitForInputIdle
